requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Testing==0.8.0
Flask-JWT-Extended==3.24.1
flask-sqlalchemy==2.4.3
flask-marshmallow==0.13.0
marshmallow==3.6.1
numpy==1.18.5
regex==2020.6.8
webargs==6.1.0
Flask-Redis==0.4.0
gunicorn==20.0.4
jsonschema==3.2.0
pymysql==0.9.3
mysqlclient==1.4.6
celery==4.4.5
marshmallow-sqlalchemy==0.23.1
pytest==5.4.2
coverage==5.1
flask-sqlacodegen==1.1.7
password-strength==0.0.3.post2
pyinstaller==3.6
pypiwin32
apscheduler==3.6.3
flask_cors
